MISUV-7388:   * more code refactoring;
