+++ app/controllers/TaxYearSummaryController.scala
@@ -36,6 +36,7 @@ import play.api.Logger
 import play.api.i18n.{I18nSupport, Lang, Messages, MessagesApi}
 import play.api.mvc._
 import services._
+import services.claimToAdjustPOA.ClaimToAdjustService
 import uk.gov.hmrc.auth.core.AuthorisedFunctions
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.play.language.LanguageUtils
