+++ app/controllers/claimToAdjustPOA/AmendablePOAController.scala
@@ -26,7 +26,7 @@ import models.core.Nino
 import play.api.Logger
 import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
-import services.ClaimToAdjustService
+import services.claimToAdjustPOA.ClaimToAdjustService
 import uk.gov.hmrc.auth.core.AuthorisedFunctions
 import utils.AuthenticatorPredicate
 import views.html.claimToAdjustPoa.AmendablePaymentOnAccount
