+++ app/services/claimToAdjustPOA/ClaimToAdjustCalculationListHelper.scala
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2024 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package services.claimToAdjustPOA
+
+import connectors.CalculationListConnector
+import models.calculationList.{CalculationListErrorModel, CalculationListModel}
+import models.core.Nino
+import models.incomeSourceDetails.TaxYear
+import play.api.http.Status.NO_CONTENT
+import services.DateServiceInterface
+import uk.gov.hmrc.http.{HeaderCarrier, InternalServerException}
+
+import scala.concurrent.{ExecutionContext, Future}
+
+trait ClaimToAdjustCalculationListHelper {
+
+  private def isTaxYearNonCrystallised(taxYear: TaxYear, nino: Nino)
+                                      (implicit hc: HeaderCarrier, dateService: DateServiceInterface,
+                                         calculationListConnector: CalculationListConnector, ec: ExecutionContext): Future[Boolean] = {
+    if (taxYear.isFutureTaxYear(dateService)) {
+      Future.successful(true)
+    } else {
+      calculationListConnector.getCalculationList(nino, taxYear.formatTaxYearRange).flatMap {
+        case res: CalculationListModel => Future.successful(res.crystallised.getOrElse(false))
+        case err: CalculationListErrorModel if err.code == NO_CONTENT => Future.successful(false)
+        case err: CalculationListErrorModel => Future.failed(new InternalServerException(err.message))
+      }.map(!_)
+    }
+  }
+
+  protected def checkCrystallisation(nino: Nino, taxYearList: List[TaxYear])
+                                    (implicit hc: HeaderCarrier, dateService: DateServiceInterface,
+                                     calculationListConnector: CalculationListConnector, ec: ExecutionContext): Future[Option[TaxYear]] = {
+    taxYearList.foldLeft(Future.successful(Option.empty[TaxYear])) { (acc, item) =>
+      acc.flatMap {
+        case Some(_) => acc
+        case None => isTaxYearNonCrystallised(item, nino)(hc, dateService, calculationListConnector, ec) map {
+          case true => Some(item)
+          case false => None
+        }
+      }
+    }
+  }
+  
+}
