+++ test/controllers/WhatYouOweControllerSpec.scala
@@ -31,8 +31,9 @@ import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.{mock, when}
 import play.api.http.Status
 import play.api.mvc.{MessagesControllerComponents, Result}
-import play.api.test.Helpers.{status, _}
-import services.{ClaimToAdjustService, WhatYouOweService}
+import play.api.test.Helpers._
+import services.WhatYouOweService
+import services.claimToAdjustPOA.ClaimToAdjustService
 import testConstants.BaseTestConstants
 import testConstants.BaseTestConstants.testAgentAuthRetrievalSuccess
 import testConstants.FinancialDetailsTestConstants._
